(MAINT) Move version.rb to work with ci
Jenkins expects the version.rb file to be under lib/<github_project>/version.rb
