+++ lib/ruby-hocon/version.rb
